+++ spec/support/shared_examples/check_digit_identifier.rb
@@ -28,8 +28,8 @@ RSpec.shared_examples 'a check-digit identifier' do |params|
         expect(instance.valid?).to be(true)
       end
 
-      it 'returns true for #check_digit_valid?' do
-        expect(instance.check_digit_valid?).to be(true)
+      it 'returns true for #valid_check_digit?' do
+        expect(instance.valid_check_digit?).to be(true)
       end
 
       it 'returns true for #valid_format?' do
@@ -53,8 +53,8 @@ RSpec.shared_examples 'a check-digit identifier' do |params|
         expect(instance.valid?).to be(false)
       end
 
-      it 'returns false for #check_digit_valid?' do
-        expect(instance.check_digit_valid?).to be(false)
+      it 'returns false for #valid_check_digit?' do
+        expect(instance.valid_check_digit?).to be(false)
       end
 
       it 'returns true for #valid_format?' do
@@ -78,8 +78,8 @@ RSpec.shared_examples 'a check-digit identifier' do |params|
         expect(instance.valid?).to be(false)
       end
 
-      it 'returns false for #check_digit_valid?' do
-        expect(instance.check_digit_valid?).to be(false)
+      it 'returns false for #valid_check_digit?' do
+        expect(instance.valid_check_digit?).to be(false)
       end
 
       it 'returns true for #valid_format?' do
@@ -99,8 +99,8 @@ RSpec.shared_examples 'a check-digit identifier' do |params|
         expect(instance.valid?).to be(false)
       end
 
-      it 'returns false for #check_digit_valid?' do
-        expect(instance.check_digit_valid?).to be(false)
+      it 'returns false for #valid_check_digit?' do
+        expect(instance.valid_check_digit?).to be(false)
       end
 
       it 'returns false for #valid_format?' do
@@ -190,21 +190,21 @@ RSpec.shared_examples 'a check-digit identifier' do |params|
       end
     end
 
-    describe '.check_digit_valid?' do
+    describe '.valid_check_digit?' do
       it 'returns true for valid identifier' do
-        expect(identifier_class.check_digit_valid?(valid_id)).to be(true)
+        expect(identifier_class.valid_check_digit?(valid_id)).to be(true)
       end
 
       it 'returns false for identifier without check-digit' do
-        expect(identifier_class.check_digit_valid?(valid_id_without_check)).to be(false)
+        expect(identifier_class.valid_check_digit?(valid_id_without_check)).to be(false)
       end
 
       it 'returns false for identifier with invalid check-digit' do
-        expect(identifier_class.check_digit_valid?(invalid_check_digit_id)).to be(false)
+        expect(identifier_class.valid_check_digit?(invalid_check_digit_id)).to be(false)
       end
 
       it 'returns false for invalid format' do
-        expect(identifier_class.check_digit_valid?(invalid_format_id)).to be(false)
+        expect(identifier_class.valid_check_digit?(invalid_format_id)).to be(false)
       end
     end
   end